Cambios realizados: • Se mejoro la interfaz del recordatorio • pestaña progreso completada • pestaña perfil mejorada

+++ app/src/main/java/com/diazmain/obapp/Home/model/MeasuresValue.java
@@ -3,41 +3,43 @@ package com.diazmain.obapp.Home.model;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigDecimal;
+
 public class MeasuresValue {
 
     @SerializedName("peso")
     @Expose
-    private double peso;
+    private BigDecimal peso;
 
     @SerializedName("cintura")
     @Expose
-    private double cintura;
+    private BigDecimal cintura;
 
     @SerializedName("grasa")
     @Expose
-    private double grasa;
+    private BigDecimal grasa;
 
-    public double getPeso() {
+    public BigDecimal getPeso() {
         return peso;
     }
 
-    public void setPeso(double peso) {
+    public void setPeso(BigDecimal peso) {
         this.peso = peso;
     }
 
-    public double getCintura() {
+    public BigDecimal getCintura() {
         return cintura;
     }
 
-    public void setCintura(double cintura) {
+    public void setCintura(BigDecimal cintura) {
         this.cintura = cintura;
     }
 
-    public double getGrasa() {
+    public BigDecimal getGrasa() {
         return grasa;
     }
 
-    public void setGrasa(double grasa) {
+    public void setGrasa(BigDecimal grasa) {
         this.grasa = grasa;
     }
 }
